Bitonic Sort
Bitonic Sort is a classic parallel algorithm for sorting.

Bitonic sort does O(n Log 2n) comparisons.
The number of comparisons done by Bitonic sort is more than popular sorting algorithms like Merge Sort [ does O(nlogn) comparisons], but Bitonice sort is better for parallel implementation because we always compare elements in a predefined sequence and the sequence of comparison doesn’t depend on data. Therefore it is suitable for implementation in hardware and parallel processor array.
Bitonic Sort must be done if the number of elements to sort is 2^n. The procedure of the bitonic sequence fails if the number of elements is not in the aforementioned quantity precisely.
It Will show the output:
Sorted array: 
1 2 3 4 5 6 7 8