new node at the beginning of the Circular Linked List
In this program, we will create a circular linked list and insert every new node at the beginning of the list. If the list is empty, then head and tail will point to the newly added node. If the list is not empty, then we will store the data of the head into a temporary node temp and make new node as the head. This new head will point to the temporary node. In simple words, the newly added node will be the first node(head) and previous head(temp) will become the second node of the list.