Create Program to check Vowel or Consonant using Switch Case
In this program I am not using break statement with cases intentionally, so that if user enters any vowel, the program continues to execute all the subsequent cases until Case 'U' is reached and thats where we are setting up the value of a boolean variable to true. This way you identify that the alphabet entered by user is vowel or not.